create seperate class for Apache tiles + delete tiles config from WebConfig.java
