+++ src/main/java/com/softserverinc/edu/config/TilesConfig.java
@@ -1,0 +1,43 @@
+package com.softserverinc.edu.config;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+import org.springframework.web.servlet.view.UrlBasedViewResolver;
+import org.springframework.web.servlet.view.tiles3.TilesConfigurer;
+import org.springframework.web.servlet.view.tiles3.TilesView;
+
+/**
+ * Created by volodymyr on 7/30/16.
+ */
+@Configuration
+public class TilesConfig extends WebMvcConfigurerAdapter {
+
+    public  static final Logger LOGGER = LoggerFactory.getLogger(TilesConfig.class);
+
+    /**
+     * Tells to the Tiles context a configuration file
+     * @return
+     */
+    @Bean
+    public TilesConfigurer tilesConfigurer(){
+        TilesConfigurer tilesConfigurer = new TilesConfigurer();
+        tilesConfigurer.setDefinitions(new String[]{"/WEB-INF/layouts/tiles-definitions.xml"});
+        tilesConfigurer.setCheckRefresh(true);
+        return tilesConfigurer;
+    }
+
+    /**
+     * Let to rosolve of views to Tiles templating framework
+     * @return
+     */
+    @Bean
+    public UrlBasedViewResolver viewResolver() {
+        UrlBasedViewResolver viewResolver = new UrlBasedViewResolver();
+        viewResolver.setViewClass(TilesView.class);
+        return viewResolver;
+    }
+
+}
